pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ashare-alpha-lab"
version = "0.1.0"
description = "A-share stock research, backtesting, signal, and reporting lab"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2",
  "pyyaml",
]

[project.scripts]
ashare-alpha = "ashare_alpha.cli:main"

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["ashare_alpha*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "tests"]

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = [
  ".tmp",
  "src/a_normal",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
